Revert "moved model, helper, and controller into lib/ so they are only loaded once in apps that use Arturo," mess with class reloading.
When FeaturesController is defined in lib/ and loaded only once, it can't take
advantage of helpers and controller methods defined in the application, which
means it's hard to use with layouts that the application uses. Also removed
"unloadable" from FeaturesController (the app can declare this if needed)
and set load_once_paths in rails/init.rb.

This reverts commit 459e82e3e9bf449f313d93a383d96d2f906b0997.