Rename feature to feature_methods, reformat
This commit renames `arturo/feature` to `arturo/feature_methods`. The
reason for this is that we expect to have an `Arturo::Feature` class
which collides with the filename of `arturo/feature` in the library.
This causes problems with `zeitwerk` and `require`.

The module we define in `arturo/feature` is actually called
`FeatureMethods` so this name change makes it more consistent with the contents.

This also changes `require` to `require_relative` and moves the require
out of the module and to the top of the file.

Releases v3.0.1

Co-authored-by: Benjamin Quorning <bquorning@zendesk.com>