load non-controller modules from outside engine
Apps that do not use the `FeaturesController` still need to load `FeatureAvailability` and `ControllerFilters` modules (and include them into other controllers). Instead of setting up a hook for loading them as part of the Rails engine, we should be doing this in the root `Arturo` file. This way, apps that don’t use the engine will still load these modules but will not run into issues with autoloading and eager loading caused by the engine.